requirements.txt
agate==1.9.1
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
backports.tarfile==1.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
daff==1.3.46
dbt-adapters==1.14.5
dbt-common==1.17.0
dbt-core==1.9.4
dbt-extractor==0.6.0
dbt-semantic-interfaces==0.7.4
dbt-snowflake==1.9.2
deepdiff==7.0.1
filelock==3.18.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.6.0
msgpack==1.1.0
networkx==3.4.2
ordered-set==4.1.0
packaging==24.2
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.3.7
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
snowflake-connector-python==3.14.0
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
sqlparse==0.5.3
text-unidecode==1.3
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
zipp==3.21.0
snowflake-snowpark-python==1.12.0
